refactor: Move Activity and Price models to domain layer and add remote response classes
