+++ core/data/src/main/java/com/projectlab/core/data/remote/ActivitiesSearchResponse.kt
@@ -1,0 +1,7 @@
+package com.projectlab.core.data.remote
+
+import com.projectlab.core.domain.model.Activity
+
+data class ActivitiesSearchResponse(
+    val data: List<Activity>
+)
