+++ core/data/src/main/java/com/projectlab/core/data/remote/ActivityDetailResponse.kt
@@ -1,0 +1,7 @@
+package com.projectlab.core.data.remote
+
+import com.projectlab.core.domain.model.Activity
+
+data class ActivityDetailResponse(
+    val data: Activity
+)
